TKO-51: finally move the users into the models folder lol
