+++ app/ModelRepositoryInterfaces/UserModelRepositoryInterface.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 namespace App\ModelRepositoryInterfaces;
 
-use App\User;
+use App\Models\User;
 
 interface UserModelRepositoryInterface
 {
