+++ app/Models/RegistrationAccessToken.php
@@ -19,6 +19,6 @@ class RegistrationAccessToken extends Model
     ];
 
     public function user() {
-        return $this->belongsTo('App\User', 'id');
+        return $this->belongsTo('App\Models\User', 'id');
     }
 }
