+++ app/Models/StudentInfo.php
@@ -28,6 +28,6 @@ class StudentInfo extends Model
     ];
 
     public function users(){
-        return $this->belongsTo('App\User', 'id', 'user_id');
+        return $this->belongsTo('App\Models\User', 'id', 'user_id');
     }
 }
